requirements.txt.txt
python-binance==1.0.19
pandas>=2.0.0
numpy>=1.24.0
backtesting>=0.3.3